requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
clickhouse-connect==0.5.23
cramjam==2.6.2
fastparquet==2023.4.0
filelock==3.12.0
fsspec==2023.4.0
ftfy==6.1.1
idna==3.4
Jinja2==3.1.2
lz4==4.3.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
py-markdown-table==0.3.3
pyarrow==11.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
regex==2023.5.5
requests==2.29.0
six==1.16.0
sympy==1.11.1
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
zstandard==0.21.0
